+++ src/hunana/datatypes/motif_classes.py
@@ -1,0 +1,7 @@
+class MotifClasses(object):
+    """ Defines the short and long name of Motifs. """
+
+    I = 'Index'
+    Ma = 'Major'
+    Mi = 'Minor'
+    U = 'Unique'