+++ src/hunana/dynamic_constants.py
@@ -1,0 +1,3 @@
+class DynamicConstants(object):
+    SEQ_DESCRIPTIONS = None
+    HEADER_REGEX = None